feat:image editor updated
