+++ src/renderer/src/components/editor/ImageEditor.tsx
@@ -1,0 +1,11 @@
+const ImageEditor = () => {
+  return (
+    <div className="flex-1 max-h-full flex flex-col overflow-auto px-6 py-3">
+      <h2 className="text-center font-serif font-semibold text-xl mt-2 mb-5">
+        Cover Image Generation
+      </h2>
+    </div>
+  );
+};
+
+export default ImageEditor;
